Fix read-char

+++ object.c
@@ -110,7 +110,7 @@ static void describe_barray(object o, struct xbarray *x)
   xbarray_adds(x, "#<");
   for (i = 0; i < o->barray.size; i++) {
     if (i != 0) xbarray_add(x, ' ');
-    xbarray_addf(x, "0x%x", (int)(o->barray.elt[i]));
+    xbarray_addf(x, "0x%x", o->barray.elt[i]);
     if (x->size > MAX_STR_LEN) return;
   }
   xbarray_add(x, '>');
